✨feat : 프로그램 종료기능(반복탈출) 구현

+++ src/commerce/Keyboard.java
@@ -5,12 +5,12 @@ import java.util.Scanner;
 public class Keyboard {
     private Scanner sc = new Scanner(System.in);
     // 숫자받기
-    public double inputNum (String message){
+    public int inputNum (String message){
         System.out.println(message);
-        while (!sc.hasNextDouble()){
+        while (!sc.hasNextInt()){
             System.out.println("숫자만 입력가능합니다.");
             sc.next();
-        } return sc.nextDouble();
+        } return sc.nextInt();
     }
     // 문자받기
     public String inputString(String message) {
